chore: don't run animation frames if the document is not visible.
This is a potential solution to a dashboard bug where the live page /
insights page would freeze if the user left it active in a background
tab for a really long time.

This is one of those things that's really hard to test and shouldn't
cause any side effects, so I think this should be merged and if we still
get reports of the problem then we know it's still an issue.