code: ensure that only one utilization data fetching operation happens at once
There was the possibility that the component could call `fetchData`
repeatedly, causing a lot of expensive requests to be sent to the core
api. Added a `fetchDataLock` property that is set to `true` when the
operation is happening, and cleared when the operation completes.

Also, added a test to verify that only one data fetching operation
should ever happen at once.