code: fix dashboard inactive request animation frame bug
When the dashboard live list view becomes inactive (ie, the tab is
no longer in the forground) and re-focused, a bunch of rafs would back
up and cause the page to take a long time to recover.

Originally, I thought that this was related to rafs firing on an
inactive tab. However, I discovered (and what I had expected before) that
by design they don't, since you can't see the page it doesn't make sense
for them to fire anyway.

In actuality, the rafs were all being created once the tab was
re-focused. This explained why all of our previous approaches weren't
working - if the raf is being created when the page is active, trying to
keep them from being created when the page is inactive doesn't work.

So, what happens now is that when a raf is fired, it looks at the last
time a raf was fired. If it was over three seconds ago (at least right
now), it will clear out all rafs that were created after the page was
re-activated. Then, afterwards, a callback will be fired in a setTimeout
that will re-instate the raf.