code: optimize the time segment filtering process
It turns out that `moment.utc('timestamp')` is really slow. To get
around this, run that calculation when fetching the data and then use
its cached value later on to do the filtering. This means that as long
as the timestamp of the event doesn't change (which I think is a fair
assumtipn) then all the moment conversion work will be frontloaded and
filtering will be really fast.