code: only draw data that fits within the foot traffic chart
This was an issue because it turns out that if data is requested for a
time period that doesn't subdivide into even intervals, then more data
than expected will be returned. Therefore, the chart should filter the
data prior to rendering it to only display data that fits within its
bounds. Furthermore, if removing a final section of data would leave a
gap, fill in that gap by adding a final point at the end of the chart on
days that aren't the current day (the current day hasn't finished yet!)